# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: telegram-accounts
    description: Telegram account connection operations
  - name: telegram-jobs
    description: Persistent Telegram check jobs and results
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "503":
          description: Database health check failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /telegram-accounts:
    get:
      operationId: listTelegramAccounts
      tags: [telegram-accounts]
      summary: List Telegram accounts
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: object
                required: [accounts]
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/TelegramAccount"
    post:
      operationId: startTelegramAccountLogin
      tags: [telegram-accounts]
      summary: Start Telegram account login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramLoginStart"
      responses:
        "201":
          description: OTP requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramLoginStartResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Account already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /telegram-accounts/{accountId}/login:
    post:
      operationId: completeTelegramAccountLogin
      tags: [telegram-accounts]
      summary: Submit Telegram OTP or two-step password
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramLoginVerify"
      responses:
        "200":
          description: Login state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /telegram-accounts/{accountId}/status:
    post:
      operationId: refreshTelegramAccountStatus
      tags: [telegram-accounts]
      summary: Refresh Telegram account connection status
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Refreshed account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramAccount"
        "404":
          $ref: "#/components/responses/NotFound"
  /telegram-accounts/{accountId}/check:
    post:
      operationId: checkTelegramAccountPhones
      tags: [telegram-accounts]
      summary: Check phone numbers with a connected Telegram account
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramCheckRequest"
      responses:
        "202":
          description: Durable Telegram check job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramCheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Account is already being used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /telegram-accounts/{accountId}:
    delete:
      operationId: deleteTelegramAccount
      tags: [telegram-accounts]
      summary: Delete Telegram account and its encrypted session
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Account still has an active durable job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /jobs:
    get:
      operationId: listTelegramJobs
      tags: [telegram-jobs]
      summary: List persisted Telegram check jobs
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: "#/components/schemas/TelegramJob"
  /jobs/{jobId}:
    get:
      operationId: getTelegramJob
      tags: [telegram-jobs]
      summary: Get a persisted Telegram check job with results
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramJobWithResults"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTelegramJob
      tags: [telegram-jobs]
      summary: Update the persisted job status
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramJobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Job transition conflicts with current worker state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteTelegramJob
      tags: [telegram-jobs]
      summary: Delete a persisted Telegram check job
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Worker cancellation requested before deletion can complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      required: [status, database]
      properties:
        status:
          type: string
        database:
          type: object
          required: [ok]
          properties:
            ok:
              type: boolean
            detail:
              type: string
              nullable: true
    Error:
      type: object
      properties:
        message:
          type: string
      required: [message]
    TelegramAccount:
      type: object
      required: [id, phoneNumber, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        displayName:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        status:
          type: string
          enum: [disconnected, awaiting_code, awaiting_2fa, connected, rate_limited, failed, disabled]
        lastCheckedAt:
          type: string
          format: date-time
          nullable: true
        lastError:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TelegramLoginStart:
      type: object
      required: [apiId, apiHash, phoneNumber]
      properties:
        apiId:
          type: string
          pattern: "^[0-9]+$"
        apiHash:
          type: string
          minLength: 16
        phoneNumber:
          type: string
          minLength: 7
    TelegramLoginVerify:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 4
        password:
          type: string
          minLength: 1
    TelegramLoginStartResponse:
      type: object
      required: [account]
      properties:
        account:
          $ref: "#/components/schemas/TelegramAccount"
        requiresPassword:
          type: boolean
    TelegramCheckRequest:
      type: object
      required: [phones]
      properties:
        phones:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            type: string
            minLength: 7
        maxAttempts:
          type: number
          minimum: 1
          maximum: 10
        minRequestInterval:
          type: number
          minimum: 0.1
          maximum: 60
        jobName:
          type: string
          minLength: 1
          maxLength: 120
        phoneRegion:
          type: string
          pattern: "^[A-Z]{2}$"
        autoResume:
          type: boolean
    TelegramCheckResult:
      type: object
      required: [phone, status, checkedAt]
      properties:
        phone:
          type: string
        status:
          type: string
          enum: [found, not_discoverable, invalid, error, rate_limited]
        username:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        telegramId:
          type: string
          nullable: true
        lastOnline:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        retryAfterSeconds:
          type: number
          nullable: true
        checkedAt:
          type: string
          format: date-time
    TelegramCheckResponse:
      type: object
      required: [accountId, jobId, status]
      properties:
        accountId:
          type: string
        jobId:
          type: string
        status:
          type: string
          enum: [queued]
    TelegramJob:
      type: object
      required: [id, accountId, name, status, total, processed, found, notDiscoverable, errors, createdAt, updatedAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [running, paused, queued, rate_limited, completed, failed, cancelled]
        total:
          type: number
        processed:
          type: number
        found:
          type: number
        notDiscoverable:
          type: number
        errors:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TelegramJobResult:
      type: object
      required: [phone, status, checkedAt]
      properties:
        phone:
          type: string
        status:
          type: string
          enum: [found, not_discoverable, invalid, error, rate_limited]
        username:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        telegramId:
          type: ["string", "null"]
        lastOnline:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        retryAfterSeconds:
          type: ["number", "null"]
        checkedAt:
          type: string
          format: date-time
    TelegramJobWithResults:
      allOf:
        - $ref: "#/components/schemas/TelegramJob"
        - type: object
          required: [results]
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/TelegramJobResult"
    TelegramJobUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [running, paused]