# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HealthStatusDatabase } from './healthStatusDatabase';

export interface HealthStatus {
  status: string;
  database: HealthStatusDatabase;
}