# lib/api-zod/src/generated/types/telegramCheckRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramCheckRequest {
  /**
     * @minItems 1
     * @maxItems 1000
     * @items.minLength 7
     */
  phones: string[];
  /**
     * @minimum 1
     * @maximum 10
     */
  maxAttempts?: number;
  /**
     * @minimum 0.1
     * @maximum 60
     */
  minRequestInterval?: number;
  /**
     * @minLength 1
     * @maxLength 120
     */
  jobName?: string;
  /** @pattern ^[A-Z]{2}$ */
  phoneRegion?: string;
  autoResume?: boolean;
}