# lib/api-zod/src/generated/types/telegramCheckResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TelegramCheckResponseStatus } from './telegramCheckResponseStatus';

export interface TelegramCheckResponse {
  accountId: string;
  jobId: string;
  status: TelegramCheckResponseStatus;
}